backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
google-cloud-secret-manager==2.20.0
google-cloud-storage==2.18.0
google-cloud-firestore==2.16.0
